el commit contiene:                         limpieza de build.gradle y el model "contact"                         creacion de path de tests
