+++ src/main/kotlin/com/thegrid/semaphore/model/Contact.kt
@@ -1,14 +1,9 @@
 package com.thegrid.semaphore.model
 
-import org.jetbrains.spek.api.Spek
-
-//import com.winterbe.expekt.expect
-//import com.winterbe.expekt.should
-//import org.jetbrains.spek.api.Spek
-
 /**
  * Created by crist on 6/7/2016.
  */
+
 data class Contact(val name: String, val age: Int, var telephones: MutableList<Telephone>) {
 
     fun addTelephone(telephone: Telephone) = telephones.add(telephone)
@@ -20,22 +15,3 @@ data class Contact(val name: String, val age: Int, var telephones: MutableList<T
 }
 
 data class Telephone(val number: String, val name: String)
-
-
-//ejemplo
-class SimpleTest: Spek({
-    given("a contact") {
-        var tels: MutableList<Telephone> = mutableListOf()
-        val tel = Telephone("12345","tel de juan")
-        val contact = Contact("juan",23,tels)
-
-        on("adding a tel number") {
-            contact.addTelephone(tel)
-            println(contact.telephones.first().number)
-
-            it("should return the tel number") {
-                assert(contact.telephones.first().number == "12445")
-            }
-        }
-    }
-})
